+++ bentoml/cli/__init__.py
@@ -20,37 +20,14 @@ from __future__ import print_function
 
 import json
 import click
-import multiprocessing
-
-import gunicorn.app.base
-from gunicorn.six import iteritems
 
 from bentoml.archive import load
 from bentoml.server import BentoAPIServer
+from bentoml.server.gunicorn_server import GunicornApplication, get_gunicorn_worker_count
 from bentoml.cli.click_utils import DefaultCommandGroup, conditional_argument
 
 from bentoml.cli.whichcraft import which
 from bentoml.cli.serverless import generate_serverless_bundle
-
-
-class GunicornApplication(gunicorn.app.base.BaseApplication):  # pylint: disable=abstract-method
-    """
-    A custom Gunicorn application.
-    """
-
-    def __init__(self, app, port, workers):
-        self.options = {'workers': workers, 'bind': '%s:%s' % ('127.0.0.1', port)}
-        self.application = app
-        super(GunicornApplication, self).__init__()
-
-    def load_config(self):
-        config = dict([(key, value) for key, value in iteritems(self.options)
-                       if key in self.cfg.settings and value is not None])
-        for key, value in iteritems(config):
-            self.cfg.set(key.lower(), value)
-
-    def load(self):
-        return self.application
 
 
 def create_bentoml_cli(installed_archive_path=None):
@@ -119,7 +96,7 @@ def create_bentoml_cli(installed_archive_path=None):
     @bentoml_cli.command()
     @conditional_argument(installed_archive_path is None, 'archive-path', type=click.STRING)
     @click.option('-p', '--port', type=click.INT, default=BentoAPIServer._DEFAULT_PORT)
-    @click.option('-w', '--workers', type=click.INT, default=(multiprocessing.cpu_count() // 2) + 1)
+    @click.option('-w', '--workers', type=click.INT, default=get_gunicorn_worker_count())
     def serve_gunicorn(port, workers, archive_path=installed_archive_path):
         """
         Start REST API gunicorn server hosting BentoService loaded from archive
